Move `ContextSpecification` to its own module
