+++ tests/cli/test_context_cli.py
@@ -1,6 +1,7 @@
 from statue.cli import statue_cli
-from statue.command_builder import CommandBuilder, ContextSpecification
+from statue.command_builder import CommandBuilder
 from statue.context import Context
+from statue.context_specification import ContextSpecification
 from tests.constants import (
     ARG1,
     ARG2,
