+++ tests/command_builder/test_command_builder_build_command.py
@@ -2,8 +2,9 @@ import pytest
 from pytest_cases import THIS_MODULE, case, parametrize_with_cases
 
 from statue.command import Command
-from statue.command_builder import CommandBuilder, ContextSpecification
+from statue.command_builder import CommandBuilder
 from statue.context import Context
+from statue.context_specification import ContextSpecification
 from statue.exceptions import InvalidCommand
 from tests.constants import (
     ARG1,
